DataTypes

Description

Documentation

data Movie

Movie, the first param is movie id, the second is movie title, and the last is release date

Constructors

Movie Int String String

data Actor

actor, the first param is actor id , the second is actor name, and the last is an arrary contains all movies the actor plays

Constructors

Actor Int String [Movie]