Main

Description

Documentation

main :: IO ()

Main fucntion of the Application, when we put "stack run normal", the application will run with normal mode, and "stack ghc -- -threaded appMain.hs && .app/Main parallel +RTS -N2" to run the bank system in two cores

step 1. spawn an account id list and excute the run function

syntaxError :: IO ()

syntaxError

run :: [Int] -> IO ()

Run the bank simulator with concrete account id list step 1, creat 10 customers step 2, simulate transferring step 3. print the current account balance after the each transferring